refactor: useHostUpdateMutation
