+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Footer/HostSidePanel/HostForm/HostTitleFieldForm/HostTitleFieldForm.spec.tsx
@@ -5,8 +5,8 @@ import { fireEvent, render, waitFor } from '@testing-library/react'
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 import { JourneyFields as Journey } from '@core/journeys/ui/JourneyProvider/__generated__/JourneyFields'
 
-import { useHostUpdate } from '../../../../../../../../../libs/useHostUpdate'
-import { UPDATE_HOST } from '../../../../../../../../../libs/useHostUpdate/useHostUpdate'
+import { useHostUpdateMutation } from '../../../../../../../../../libs/useHostUpdateMutation'
+import { UPDATE_HOST } from '../../../../../../../../../libs/useHostUpdateMutation/useHostUpdateMutation'
 
 import {
   CREATE_HOST,
@@ -19,13 +19,13 @@ jest.mock('uuid', () => ({
   v4: jest.fn()
 }))
 
-jest.mock('../../../../../../../../../libs/useHostUpdate', () => ({
+jest.mock('../../../../../../../../../libs/useHostUpdateMutation', () => ({
   __esModule: true,
-  useHostUpdate: jest.fn()
+  useHostUpdateMutation: jest.fn()
 }))
 
-const mockUseHostUpdate = useHostUpdate as jest.MockedFunction<
-  typeof useHostUpdate
+const mockUseHostUpdate = useHostUpdateMutation as jest.MockedFunction<
+  typeof useHostUpdateMutation
 >
 
 describe('HostTitleFieldForm', () => {
