--- apps/journeys-admin/src/libs/useHostUpdate/index.ts
@@ -1,1 +1,0 @@
-export { useHostUpdate, UPDATE_HOST } from './useHostUpdate'
